src/backend/unified_service/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pyjwt==2.8.0
python-multipart==0.0.6
boto3==1.34.0